requirements.txt
langchain
langchain-google-genai
python-dotenv
streamlit
chromadb
pypdf
ragas
beautifulsoup4